review(node/sync): split mock into sibling file + add assert_all_consumed helper
Addresses 4 findings on PR #2408 (review-4325491371):

1. **Split `network.rs` into `network/mod.rs` + `network/mock.rs`**.
   Matches the `manager/mod.rs` + `manager/tests.rs` precedent and
   the workspace convention to keep test infrastructure out of the
   production module file. The mock was already `#[cfg(test)]`-gated
   so this is purely a structural cleanup — no production-binary
   surface change.

2. **Drop fully-qualified syntax in the blanket impl**. Rust's
   method resolution prefers inherent methods over trait methods,
   so `self.mesh_peers(topic).await` inside
   `impl SyncNetwork for NetworkClient` dispatches to
   `NetworkClient::mesh_peers` — no infinite recursion. Verified by
   compile + test pass. Added an inline comment so future readers
   don't reach for the qualifier defensively.

3. **Add `MockSyncNetwork::assert_all_consumed()`**. Flagged in two
   successive reviews — addresses the "test queues more responses
   than the code under test consumes, silently passes" gap. Accepts
   ≤1 leftover `mesh_peers` entry (sticky-last semantic) and 0
   leftover `open_stream` entries (each is a distinct attempt).
   `#[track_caller]` so panics point at the assertion site, not
   inside the helper. 4 new tests cover the happy paths and the
   two panic paths.

4. **Document the exhaustion asymmetry at module level**. The "sticky
   last" vs "error on exhaust" split was only mentioned in inline
   comments before; now it's an "Exhaustion semantics" section in
   the module doc explaining *why* the two behaviours differ
   (production analogue: mesh stable after discovery vs each
   open_stream is a distinct attempt).

10 mock tests pass (6 original + 4 new for the consumption helper);
162 total node tests pass; fmt + clippy clean.
